refactor(compositor): one LayerCB, one LiveParams, for both backends
Four items existed twice, once per backend, and the duplication had already done
what duplication does.

**LayerCB / Layer.** The same ten fields in the same order, declared in both engines.
They are the Rust half of a three-way contract — `cbuffer Layer` in shaders.hlsl,
`struct Layer` in shaders.metal, this struct — and a mismatch does not raise
anything, it makes a shader read `color` where `fx` was written. Unified on the
macOS spelling (`repr(C, align(16))`); under plain `repr(C)` both already produced
offsets 0/16/32/40/44/48/64/80/96/112, and Windows only `copy_nonoverlapping`s the
struct into a mapped constant buffer, where source alignment is irrelevant. The two
forms were compatible; unifying them stops them from ceasing to be. The size
assertion moves to the shared module and gains a per-field `offset_of!` check.

The doc said "64 octets". It was never right — ten fields, thirty-two `f32`.

**LiveParams and friends.** The macOS copies were annotated "mêmes champs et même
layout" and "mêmes formules". Both claims were false:

    bg_color default        windows [0.10, 0.11, 0.14, 1.0]   macos [0, 0, 0, 0]
    has_webcam default      windows true                       macos false
    webcam_shape_code(_)    windows 3 ("rounded")              macos 0 ("rectangle")

The third is the sharp one. `live_params_from_scene` calls it and the app's default
`webcam_shape` is "rounded", so one scene described a rounded camera on Windows and
a rectangular one on macOS — a silent cross-platform rendering difference sitting in
the code that exists to prevent cross-platform rendering differences.

Windows values win: that is the backend that renders in production. This changes
macOS behaviour by exactly nothing today, because the Metal engine reads none of
these fields yet — `compose_frame` hardcodes its LayerCB. Which is why it is worth
doing now: after the layer driving lands, the same fix is a regression hunt.

`OUT_W` / `OUT_H` / `FIXTURE_FRAMES` were duplicated too, with identical values.
Both compositor modules `pub use` everything back out, so `live.rs`,
`pipeline_windows.rs`, `compositor-view-napi` and `crates/poc-d3d` are untouched.
